ml/requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
pillow>=10.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
kaggle>=1.5.16
